pyproject.toml
[project]
name = "matchamuse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "groq>=0.31.1",
    "python-dotenv>=1.1.1",
]
